queueTrigger_func/requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
azure-storage-blob
azure-storage-queue
pydantic
